M1 Stats

+++ Scripts/Physiology/findBins.m
@@ -2,6 +2,6 @@ function bins = findBins(timesT,allBins)
 if(~any(~isnan(timesT)))
     bins =NaN(size(timesT));
 else
-    bins = int64(discretize(timesT,allBins));
+    bins = discretize(timesT,allBins);
 end
 end
